requirements.txt
chainlit
langchain
google-generativeai
langchain-google-genai